requirements.txt
# Requirements for all examples in the book.
# Tested with Python version 3.10.14
bertviz==1.4.0
faiss-cpu==1.8.0.post1
ipython==8.26.0
keras==3.4.1
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-experimental==0.0.62
langchain-text-splitters==0.2.2
matplotlib==3.9.1
nltk==3.8.1
numpy==1.26.4
# You may change the version to match the version of CUDA installed on your computer.
nvidia-cudnn-cu11==8.9.6.50
openai==1.35.10
pandas==2.2.2
protobuf==3.20.3
pyspellchecker==0.8.1
regex==2024.5.15
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
tabulate==0.9.0
tensorflow[and-cuda]==2.16.2
tensorflow-datasets==4.9.6
tensorflow-estimator==2.15.0
transformers==4.42.3
web-browser==0.0.1